pyproject.toml
[project]
name = "fufufafaethon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-whisper",
    "yt-dlp",
]

[tool.ruff]
line-length = 120

[tool.uv.sources]
openai-whisper = { git = "https://github.com/openai/whisper.git" }
yt-dlp = { git = "https://github.com/yt-dlp/yt-dlp" }
